% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloader.R
\name{load_api}
\alias{load_api}
\title{Load Data from API (JSON response)}
\usage{
load_api(url)
}
\arguments{
\item{url}{A string specifying the URL of the API.}
}
\value{
A data frame containing the parsed JSON response, or \code{NULL} in case of an error.
}
\description{
This function sends a GET request to a specified API endpoint and returns the response as a data frame.
It expects the response to be in JSON format.
}
\examples{
\dontrun{
load_api("https://api.example.com/data")
}
}
