% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloader.R
\name{load_csv}
\alias{load_csv}
\title{Load CSV File}
\usage{
load_csv(file_path)
}
\arguments{
\item{file_path}{The path to the CSV file.}
}
\value{
A data frame containing the contents of the CSV file.
}
\description{
This function loads a CSV file from the specified file path and returns the data as a data frame.
}
\examples{
\dontrun{
load_csv("data/sample.csv")
}
}
