% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloader.R
\name{load_data}
\alias{load_data}
\title{Automate Data Loading from Various Sources}
\usage{
load_data(source, source_type, ...)
}
\arguments{
\item{source}{A string specifying the path or URL of the data source.}

\item{source_type}{A string specifying the type of the data source.}

\item{...}{Additional arguments to pass to the underlying loader functions.}
}
\value{
The loaded data or \code{NULL} in case of an error.
}
\description{
Automate Data Loading from Various Sources
}
