% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloader.R
\name{load_excel}
\alias{load_excel}
\title{Load Excel File}
\usage{
load_excel(file_path, sheet = 1)
}
\arguments{
\item{file_path}{A string specifying the path to the Excel file.}

\item{sheet}{An optional integer specifying the sheet number to load. Default is 1.}
}
\value{
A data frame containing the contents of the specified Excel sheet, or \code{NULL} in case of an error.
}
\description{
This function loads an Excel file from the specified file path and returns the data as a data frame.
It allows specifying a particular sheet and handles errors by returning \code{NULL}.
}
\examples{
\dontrun{
load_excel("data/sample.xlsx", sheet = 1)
}
}
