% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloader.R
\name{load_html_table}
\alias{load_html_table}
\title{Load HTML Table from a Website}
\usage{
load_html_table(url, table_num = 1)
}
\arguments{
\item{url}{A string specifying the URL of the website.}

\item{table_num}{An optional integer specifying which table to extract. Default is 1.}
}
\value{
A data frame containing the contents of the specified HTML table, or \code{NULL} in case of an error.
}
\description{
This function loads an HTML page from the specified URL and extracts a table from it.
The function allows you to specify which table to return if the page contains multiple tables.
}
\examples{
\dontrun{
load_html_table("https://example.com/page", table_num = 1)
}
}
