% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataloader.R
\name{load_pdf}
\alias{load_pdf}
\title{Load Data from a PDF File}
\usage{
load_pdf(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the PDF file.}
}
\value{
A character vector containing the text from the PDF file, or \code{NULL} in case of an error.
}
\description{
This function extracts the text from a PDF file and returns it as a character vector.
Each element of the vector corresponds to a page in the PDF.
}
\examples{
\dontrun{
pdf_text <- load_pdf("data/sample.pdf")
}
}
